requirements.txt
requests
Pillow
PyMuPDF
watchdog
